Implemented a gameOver variable.
Also cleaned up some comments and shortened the Event Listeners by using arrow functions and implementing logic inside of them so that they are only activated if the conditions are met. This also allowed me to pass the event functions in with parameters instead of having to define them as named functions seperately and passing them in as variables.